+++ OpenBudgeteer.Blazor/ViewModels/ApexReportViewModel.cs
@@ -11,13 +11,20 @@ namespace OpenBudgeteer.Blazor.ViewModels;
 
 public class ApexReportViewModel : ReportPageViewModel
 {
-    public List<ReportRecord> MonthBalances { get; private set; } = new();
-    public List<ReportRecord> BankBalances { get; private set; } = new();
-    public List<ReportRecord> MonthIncome { get; private set; } = new();
-    public List<ReportRecord> MonthExpenses { get; private set; } = new();
-    public List<ReportRecord> YearIncome { get; private set; } = new();
-    public List<ReportRecord> YearExpenses { get; private set; } = new();
-    public List<Tuple<string, List<ReportRecord>>> MonthBucketExpenses { get; private set; } = new();
+    public List<ApexRecord> MonthBalances { get; private set; } = [];
+    public List<ApexRecord> BankBalances { get; private set; } = [];
+    public List<ApexRecord> MonthIncome { get; private set; } = [];
+    public List<ApexRecord> MonthExpenses { get; private set; } = [];
+    public List<ApexRecord> YearIncome { get; private set; } = [];
+    public List<ApexRecord> YearExpenses { get; private set; } = [];
+    public List<ApexRecord> BalanceDistributionBucketGroup { get; private set; } = [];
+    public List<ApexRecord> BalanceDistributionBucket { get; private set; } = [];
+    public List<ApexRecord> InDistributionBucketGroup { get; private set; } = [];
+    public List<ApexRecord> InDistributionBucket { get; private set; } = [];
+    public List<ApexRecord> ActivityDistributionBucketGroup { get; private set; } = [];
+    public List<ApexRecord> ActivityDistributionBucket { get; private set; } = [];
+    public List<BucketGroupApexRecord> BucketGroupsBucketBudgets { get; private set; } = [];
+    public List<Tuple<string, List<ApexRecord>>> MonthBucketExpenses { get; private set; } = [];
     
     private readonly AppSettingService _appSettingService;
     
@@ -28,13 +35,14 @@ public class ApexReportViewModel : ReportPageViewModel
     
     public async Task LoadDataAsync()
     {
-        var loadTasks = new List<Task>()
+        var loadTasks = new List<Task>
         {
             LoadMonthBalancesReportAsync(_appSettingService.CurrentSettings.ReportCounts.MonthBalance),
             LoadBankBalancesReportAsync(_appSettingService.CurrentSettings.ReportCounts.BankBalance),
             LoadMonthIncomeExpensesReportAsync(_appSettingService.CurrentSettings.ReportCounts.MonthIncomeExpenses),
             LoadYearIncomeExpensesReportAsync(_appSettingService.CurrentSettings.ReportCounts.YearIncomeExpenses),
-            LoadMonthExpensesBucketReportAsync(_appSettingService.CurrentSettings.ReportCounts.MonthBucketExpenses)
+            LoadMonthExpensesBucketReportAsync(_appSettingService.CurrentSettings.ReportCounts.MonthBucketExpenses),
+            LoadBucketReportsAsync()
         };
         await Task.WhenAll(loadTasks);
     }
@@ -44,7 +52,7 @@ public class ApexReportViewModel : ReportPageViewModel
         MonthBalances.Clear();
         foreach (var (month, balance) in await LoadMonthBalancesAsync(months))
         {
-            MonthBalances.Add(new ReportRecord(month.ToString("yyyy-MM"), balance));
+            MonthBalances.Add(new ApexRecord(month.ToString("yyyy-MM"), balance));
         }
     }
 
@@ -53,7 +61,7 @@ public class ApexReportViewModel : ReportPageViewModel
         BankBalances.Clear();
         foreach (var (month, balance) in await LoadBankBalancesAsync(months))
         {
-            BankBalances.Add(new ReportRecord(month.ToString("yyyy-MM"), balance));
+            BankBalances.Add(new ApexRecord(month.ToString("yyyy-MM"), balance));
         }
     }
     
@@ -63,8 +71,8 @@ public class ApexReportViewModel : ReportPageViewModel
         MonthExpenses.Clear();
         foreach (var (month, income, expenses) in await LoadMonthIncomeExpensesAsync(months))
         {
-            MonthIncome.Add(new ReportRecord(month.ToString("yyyy-MM"), income));
-            MonthExpenses.Add(new ReportRecord(month.ToString("yyyy-MM"), expenses));
+            MonthIncome.Add(new ApexRecord(month.ToString("yyyy-MM"), income));
+            MonthExpenses.Add(new ApexRecord(month.ToString("yyyy-MM"), expenses));
         }
     }
     
@@ -74,8 +82,8 @@ public class ApexReportViewModel : ReportPageViewModel
         YearExpenses.Clear();
         foreach (var (year, income, expenses) in await LoadYearIncomeExpensesAsync(years))
         {
-            YearIncome.Add(new ReportRecord(year.ToString("yyyy"), income));
-            YearExpenses.Add(new ReportRecord(year.ToString("yyyy"), expenses));
+            YearIncome.Add(new ApexRecord(year.ToString("yyyy"), income));
+            YearExpenses.Add(new ApexRecord(year.ToString("yyyy"), expenses));
         }
     }
     
@@ -85,9 +93,52 @@ public class ApexReportViewModel : ReportPageViewModel
         foreach (var item in await LoadMonthExpensesBucketAsync(months))
         {
             var data = item.MonthlyResults
-                .Select(i => new ReportRecord(i.Item1.ToString("yyyy-MM"), i.Item2))
+                .Select(i => new ApexRecord(i.Item1.ToString("yyyy-MM"), i.Item2))
                 .ToList();
             MonthBucketExpenses.Add(new(item.BucketName, data));
+        }
+    }
+
+    private async Task LoadBucketReportsAsync()
+    {
+        BalanceDistributionBucketGroup.Clear();
+        BalanceDistributionBucket.Clear();
+        InDistributionBucketGroup.Clear();
+        InDistributionBucket.Clear();
+        ActivityDistributionBucketGroup.Clear();
+        ActivityDistributionBucket.Clear();
+        BucketGroupsBucketBudgets.Clear();
+
+        var bucketReportResult = await LoadBucketStatisticsAsync();
+
+        BalanceDistributionBucketGroup.AddRange(bucketReportResult.BalancesPerBucketGroup
+            .Select(i => new ApexRecord(i.Item1, i.Item2)));
+        BalanceDistributionBucket.AddRange(bucketReportResult.BalancesPerBucket
+            .Select(i => new ApexRecord(i.Item1, i.Item2)));
+        InDistributionBucketGroup.AddRange(bucketReportResult.InPerBucketGroup
+            .Select(i => new ApexRecord(i.Item1, i.Item2)));
+        InDistributionBucket.AddRange(bucketReportResult.InPerBucket
+            .Select(i => new ApexRecord(i.Item1, i.Item2)));
+        ActivityDistributionBucketGroup.AddRange(bucketReportResult.ActivityPerBucketGroup
+            .Select(i => new ApexRecord(i.Item1, i.Item2 * -1)));
+        ActivityDistributionBucket.AddRange(bucketReportResult.ActivityPerBucket
+            .Select(i => new ApexRecord(i.Item1, i.Item2 * -1)));
+
+        foreach (var (bucketGroup, buckets) in bucketReportResult.BudgetConsumptionPerBucket)
+        {
+            BucketGroupsBucketBudgets.Add(new(
+                bucketGroup, 
+                buckets
+                    .Select(i => new BucketApexRecord(i.Item1, ConvertToGaugeItemSource(i)))
+                    .ToList()
+                )
+            );
+        }
+        return;
+
+        List<ApexRecord> ConvertToGaugeItemSource(Tuple<string, decimal> source)
+        {
+            return [new ApexRecord("Remaining", Math.Round(source.Item2, 0))];
         }
     }
 }